requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
pymysql
python-dotenv
passlib[bcrypt]
PyJWT
